flutter final project/backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
Pillow==11.0.0
python-dotenv==1.0.1
dj-rest-auth==5.0.2
django-allauth==0.61.1
djangorestframework-simplejwt==5.3.1 